# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference : TMF 638 - Service Inventory Management v4 Release\
    \ 18.5 - Dec 2018\n\nThe intent of this API is to provide a consistent/standardized\
    \ mechanism to query and manipulate the Service inventory.\n\nThe Service Inventory\
    \ API can be used to query the service instances for a customer via Self Service\
    \ Portal or the Call Centre operator can query the service instances on behalf\
    \ of the customer while a customer may have a complaint or a query.\nNote: Only\
    \ the CustomerFacingServices instances will be presented to the customer.\n\n\
    The Service Inventory API can be called by the Service Order Management to create\
    \ a new service instance/ update an existing service instance in the Service Inventory.\n\
    \n### Service resource\nService is an abstract base class for defining the Service\
    \ hierarchy. All Services are characterized as either being possibly visible and\
    \ usable by a Customer or not. This gives rise to the two subclasses of  Service:\
    \ CustomerFacingService and ResourceFacingService.\n\n### Service Inventory API\
    \ performs the following operations on service  :\n- Retrieve  a service or a\
    \ collection of services depending on filter criteria\n- Partial update of a service\
    \ (including updating rules)\n- Create a service (including default values and\
...
    \ users only)\n- Notification of events on service\n\nCopyright (c)TM Forum 2018.\
...
host: "localhost:8080"
basePath: "/tmf-api"
...
schemes:
...
consumes:
...
produces:
...
paths:
...
  /listener/serviceCreateNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceCreateNotification"
      description: "Example of a client listener for receiving the notification ServiceCreateNotification"
      operationId: "listenToServiceCreateNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceCreateNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceAttributeValueChangeNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceAttributeValueChangeNotification"
      description: "Example of a client listener for receiving the notification ServiceAttributeValueChangeNotification"
      operationId: "listenToServiceAttributeValueChangeNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceAttributeValueChangeNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceStateChangeNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceStateChangeNotification"
      description: "Example of a client listener for receiving the notification ServiceStateChangeNotification"
      operationId: "listenToServiceStateChangeNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceStateChangeNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceBatchNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceBatchNotification"
      description: "Example of a client listener for receiving the notification ServiceBatchNotification"
      operationId: "listenToServiceBatchNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceBatchNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceDeleteNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceDeleteNotification"
      description: "Example of a client listener for receiving the notification ServiceDeleteNotification"
      operationId: "listenToServiceDeleteNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceDeleteNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
definitions:
...
  EventSubscription:
    type: "object"
    required:
    - "callback"
    - "id"
    properties:
      id:
        type: "string"
        description: "Id of the listener"
      callback:
        type: "string"
        description: "The callback being registered."
      query:
        type: "string"
        description: "additional data to be passed"
    description: "Sets the communication endpoint address the service instance must\
      \ use to deliver notification information"
    example:
      query: "query"
      callback: "callback"
      id: "id"
...
  ServiceCreateNotification:
    type: "object"
    properties:
      eventId:
        type: "string"
        description: "The identifier of the notification"
      eventTime:
        type: "string"
        format: "date-time"
        description: "Time of the event occurence"
      eventType:
        type: "string"
        description: "The type of the notification"
      fieldPath:
        type: "string"
        description: "The path identifying the object field concerned by this notification"
      resourcePath:
        type: "string"
        description: "The path identifying the resource object concerned by this notification"
      event:
        description: "The event linked to the involved resource object"
        $ref: "#/definitions/ServiceCreateEvent"
    description: "The notification data structure"
    example:
      eventId: "eventId"
      resourcePath: "resourcePath"
      eventTime: "2000-01-23T04:56:07.000+00:00"
      fieldPath: "fieldPath"
      eventType: "eventType"
      event:
        service:
          serviceType: "serviceType"
          note:
          - date: "2000-01-23T04:56:07.000+00:00"
            system: "system"
            '@baseType': "@baseType"
            author: "author"
            '@type': "@type"
            text: "text"
            '@schemaLocation': "http://example.com/aeiou"
          - date: "2000-01-23T04:56:07.000+00:00"
            system: "system"
            '@baseType': "@baseType"
            author: "author"
            '@type': "@type"
            text: "text"
            '@schemaLocation': "http://example.com/aeiou"
          serviceDate: "serviceDate"
          endDate: "2000-01-23T04:56:07.000+00:00"
          supportingResource:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          '@type': "@type"
          hasStarted: true
          description: "description"
          serviceOrder:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            serviceOrderItemId: "serviceOrderItemId"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            serviceOrderItemId: "serviceOrderItemId"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          serviceRelationship:
          - relationshipType: "relationshipType"
            '@baseType': "@baseType"
            service:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
            '@type': "@type"
            '@schemaLocation': "http://example.com/aeiou"
          - relationshipType: "relationshipType"
            '@baseType': "@baseType"
            service:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
            '@type': "@type"
            '@schemaLocation': "http://example.com/aeiou"
          '@baseType': "@baseType"
          id: "id"
          href: "href"
          place:
          - role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          state: {}
          '@schemaLocation': "http://example.com/aeiou"
          isStateful: true
          serviceCharacteristic:
          - '@baseType': "@baseType"
            '@type': "@type"
            valueType: "valueType"
            name: "name"
            '@schemaLocation': "http://example.com/aeiou"
            value: {}
          - '@baseType': "@baseType"
            '@type': "@type"
            valueType: "valueType"
            name: "name"
            '@schemaLocation': "http://example.com/aeiou"
            value: {}
          serviceSpecification:
            targetServiceSchema:
              '@baseType': "@baseType"
              '@type': "@type"
              '@schemaLocation': "@schemaLocation"
            '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
            version: "version"
          supportingService:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          relatedParty:
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          isServiceEnabled: true
          startMode: "startMode"
          name: "name"
          category: "category"
          startDate: "2000-01-23T04:56:07.000+00:00"
...
  ServiceAttributeValueChangeNotification:
    type: "object"
    properties:
      eventId:
        type: "string"
        description: "The identifier of the notification"
      eventTime:
        type: "string"
        format: "date-time"
        description: "Time of the event occurence"
      eventType:
        type: "string"
        description: "The type of the notification"
      fieldPath:
        type: "string"
        description: "The path identifying the object field concerned by this notification"
      resourcePath:
        type: "string"
        description: "The path identifying the resource object concerned by this notification"
      event:
        description: "The event linked to the involved resource object"
        $ref: "#/definitions/ServiceAttributeValueChangeEvent"
    description: "The notification data structure"
    example:
      eventId: "eventId"
      resourcePath: "resourcePath"
      eventTime: "2000-01-23T04:56:07.000+00:00"
      fieldPath: "fieldPath"
      eventType: "eventType"
      event:
        service:
          serviceType: "serviceType"
          note:
          - date: "2000-01-23T04:56:07.000+00:00"
            system: "system"
            '@baseType': "@baseType"
            author: "author"
            '@type': "@type"
            text: "text"
            '@schemaLocation': "http://example.com/aeiou"
          - date: "2000-01-23T04:56:07.000+00:00"
            system: "system"
            '@baseType': "@baseType"
            author: "author"
            '@type': "@type"
            text: "text"
            '@schemaLocation': "http://example.com/aeiou"
          serviceDate: "serviceDate"
          endDate: "2000-01-23T04:56:07.000+00:00"
          supportingResource:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          '@type': "@type"
          hasStarted: true
          description: "description"
          serviceOrder:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            serviceOrderItemId: "serviceOrderItemId"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            serviceOrderItemId: "serviceOrderItemId"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          serviceRelationship:
          - relationshipType: "relationshipType"
            '@baseType': "@baseType"
            service:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
            '@type': "@type"
            '@schemaLocation': "http://example.com/aeiou"
          - relationshipType: "relationshipType"
            '@baseType': "@baseType"
            service:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
            '@type': "@type"
            '@schemaLocation': "http://example.com/aeiou"
          '@baseType': "@baseType"
          id: "id"
          href: "href"
          place:
          - role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          state: {}
          '@schemaLocation': "http://example.com/aeiou"
          isStateful: true
          serviceCharacteristic:
          - '@baseType': "@baseType"
            '@type': "@type"
            valueType: "valueType"
            name: "name"
            '@schemaLocation': "http://example.com/aeiou"
            value: {}
          - '@baseType': "@baseType"
            '@type': "@type"
            valueType: "valueType"
            name: "name"
            '@schemaLocation': "http://example.com/aeiou"
            value: {}
          serviceSpecification:
            targetServiceSchema:
              '@baseType': "@baseType"
              '@type': "@type"
              '@schemaLocation': "@schemaLocation"
            '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
            version: "version"
          supportingService:
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          - '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
          relatedParty:
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          isServiceEnabled: true
          startMode: "startMode"
          name: "name"
          category: "category"
          startDate: "2000-01-23T04:56:07.000+00:00"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "integer"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "integer"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"